nob/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "GlocalText"
version = "1.2.1"
description = "A CLI tool for automated i18n and l10n of software projects."
authors = [{ name = "OG-Open-Source" }]
requires-python = ">=3.9"
dependencies = [
    "pydantic",
    "pyyaml",
    "xxhash",
    "google-genai",
    "deep-translator",
    "openai",
    "requests",
]

[project.scripts]
glocaltext = "glocaltext.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]
